Optionally use the environment for config instead of properties files
Using ENV vars is more conducive to running in Docker.  If we need to run multiple streams in the same app, then properties files will need to be used.

If we want to run a single stream, it's now possible to use ENV variables for our config instead of specifying properties.  This allows us to build the code, and execute with different config in multiple docker containers.